requirements.txt
Django==2.2
djangorestframework==3.12.2
Pillow==8.1.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
